refactor + eslint-disable: move spawnProcess to check-links.js
